Fix some artifacts of angular normal map
First, for sampling angular normal map (anmap.png), the angle vaule
has distinct gap at -PI and PI. So, using linear filtering will get
an average value which is not correct.

Second, for sampling light map (light.png), the values at edges of T = 0
and T = 1 are not continous. So, using "repeat" wrap is not correct.